Cleaned up and imporved speed
