+++ src/pages/index.tsx
@@ -1,30 +1,15 @@
-import { Button, DatePicker, Flex, Select, Modal, Input, message, Form } from "antd";
+import { Flex } from "antd";
 import dayjs from "dayjs";
 import type { GetServerSideProps, InferGetServerSidePropsType } from "next";
 import Head from "next/head";
-import Image from "next/image";
-import Link from "next/link";
 import { useRouter } from "next/router";
 import { useCallback, useEffect, useState } from "react";
+import { FeedbackModal } from "~/components/FeedbackModal";
+import { Footer } from "~/components/Footer";
+import { Header } from "~/components/Header";
+import { TrainSelector } from "~/components/TrainSelector";
 import { getBaseURL, isInMaintenance } from "~/lib/deployment";
 import { getInitialTrains } from "~/lib/digitraffic";
-
-const pickerStyle: React.CSSProperties = {
-  width: "100%",
-  height: "40px",
-  maxWidth: "300px",
-};
-
-const slogans = [
-  '"Ilman vieruskaveria"',
-  '"Koska tässä kestää..."',
-  '"Noniin ja Kokkolasta lähdettiin noin 45 minuuttia myöhässä. Syynä oli junakohtaukset ja rautatieongelmat"',
-  '"Vieläkin Ratapihalla"',
-  '"Vertauskelvoton Rautaromu"',
-  '"Junan suunnan vaihtamisessa on ongelmia, jarrut ovat liian tiukalla"',
-  '"Omalla matkalla"',
-  '"Tarkoituksella yksin"',
-];
 
 export default function Home({
   initialDate,
@@ -32,19 +17,14 @@ export default function Home({
   maintenance,
 }: InferGetServerSidePropsType<typeof getServerSideProps>) {
   const router = useRouter();
-  const [sloganText, setSloganText] = useState<string | null>(null);
 
   const [allTrains, setAllTrains] = useState(initialTrains);
   const [trainsLoaded, setTrainsLoaded] = useState<boolean>(true);
-  const [trainLoading, setTrainLoading] = useState<boolean>(false);
+
+  const [isFbModalOpen, setIsFbModalOpen] = useState<boolean>(false);
 
   const [selectedDate, setSelectedDate] = useState<string | null>(initialDate);
   const [selectedTrain, setSelectedTrain] = useState<string | null>(null);
-
-  const [messageApi, messageContextHolder] = message.useMessage();
-
-  const [fbForm] = Form.useForm();
-  const [isFbModalOpen, setIsFbModalOpen] = useState<boolean>(false);
 
   useEffect(() => {
     // @ts-expect-error no types for globally available plausible function
@@ -69,11 +49,6 @@ export default function Home({
     if (!selectedDate) return;
     getTrains(selectedDate).catch(console.error);
   }, [selectedDate, getTrains, initialDate, initialTrains]);
-
-  useEffect(() => {
-    const text = slogans[Math.floor(Math.random() * slogans.length)];
-    setSloganText(text ? text : "");
-  }, []);
 
   useEffect(() => {
     if (selectedDate && selectedTrain) {
@@ -110,97 +85,7 @@ export default function Home({
         <link rel="canonical" href={getBaseURL() + "/"} />
       </Head>
 
-      {messageContextHolder}
-
-      {/* Feedback modal */}
-      <Modal
-        title="Palautelaatikko"
-        open={isFbModalOpen}
-        onOk={() => {
-          setIsFbModalOpen(false);
-          fbForm.resetFields();
-        }}
-        onCancel={() => {
-          setIsFbModalOpen(false);
-          fbForm.resetFields();
-        }}
-        centered={true}
-        footer={[
-          <Button
-            key="back"
-            onClick={() => {
-              setIsFbModalOpen(false);
-              fbForm.resetFields();
-            }}
-          >
-            Peruuta
-          </Button>,
-          <Button
-            form="myForm"
-            key="submit"
-            htmlType="submit"
-            onClick={() => {
-              fbForm.submit();
-            }}
-            type="primary"
-          >
-            Lähetä
-          </Button>,
-        ]}
-      >
-        <Form
-          form={fbForm}
-          layout="vertical"
-          style={{ marginTop: "2em" }}
-          onFinish={(values) => {
-            void (async () => {
-              setIsFbModalOpen(false);
-              fbForm.resetFields();
-              messageApi
-                .open({
-                  type: "success",
-                  content: "Kiitos palautteesta!",
-                })
-                .then(
-                  () => null,
-                  () => null
-                );
-
-              await fetch("/api/sendFeedback", {
-                method: "POST",
-                body: JSON.stringify(values),
-              });
-            })();
-          }}
-        >
-          <Form.Item
-            name="email"
-            label="Sähköposti (Vapaaehtoinen)"
-            rules={[
-              {
-                type: "email",
-                message: "Syötä kelvollinen sähköpostiosoite!",
-              },
-            ]}
-          >
-            <Input />
-          </Form.Item>
-
-          <Form.Item
-            label="Palaute"
-            name="feedback"
-            rules={[
-              { required: true, message: "Tyhjää palautetta ei voi lähettää!" },
-              { max: 1500, message: "Maksimipituus on 1500 merkkiä!" },
-            ]}
-          >
-            <Input.TextArea
-              placeholder="Anna palautetta tai kehitysideoita..."
-              autoSize={{ minRows: 3, maxRows: 5 }}
-            />
-          </Form.Item>
-        </Form>
-      </Modal>
+      <FeedbackModal isFbModalOpen={isFbModalOpen} setIsFbModalOpen={setIsFbModalOpen} />
 
       <Flex
         style={{ width: "100%", gap: "5px" }}
@@ -208,161 +93,20 @@ export default function Home({
         align={"center"}
         vertical={true}
       >
-        <div style={{ width: "calc(250px + 10vw)" }}>
-          <Link href="/" style={{ color: "unset" }}>
-            <Image
-              style={{ width: "100%", height: "auto" }}
-              src="/vr_logo.png"
-              width={475}
-              height={70}
-              alt="VenaaRauhassa"
-            />
-          </Link>
-        </div>
-        <p
-          style={{
-            fontSize: "14px",
-            fontStyle: "italic",
-            marginTop: "0px",
-            textAlign: "center",
-          }}
-        >
-          {sloganText}
-        </p>
-        <br />
+        <Header maintenance={maintenance} />
 
-        {maintenance && (
-          <h2
-            style={{
-              textAlign: "center",
-            }}
-          >
-            Palvelu on huoltokatkolla VR:n tekemien rajapintamuutoksien takia.
-            <br />
-            Asiaa selvitellään
-          </h2>
-        )}
-
-        <DatePicker
-          placeholder="Valitse päivä"
-          disabled={maintenance}
-          disabledDate={(current) => current && current < dayjs().subtract(2, "day")}
-          defaultValue={dayjs(initialDate)}
-          style={pickerStyle}
-          onChange={(_date, dateString) => {
-            if (dateString) {
-              setSelectedDate(dateString.split(".").reverse().join("-"));
-            } else {
-              setSelectedDate(null);
-              setSelectedTrain(null);
-            }
-          }}
-          format="DD.MM.YYYY"
+        <TrainSelector
+          maintenance={maintenance}
+          initialDate={initialDate}
+          selectedDate={selectedDate}
+          setSelectedDate={setSelectedDate}
+          selectedTrain={selectedTrain}
+          setSelectedTrain={setSelectedTrain}
+          trainsLoaded={trainsLoaded}
+          allTrains={allTrains}
         />
 
-        <br />
-        <Select
-          style={pickerStyle}
-          disabled={maintenance || (selectedDate && trainsLoaded ? false : true)}
-          loading={selectedDate && !trainsLoaded ? true : false}
-          showSearch
-          allowClear
-          value={selectedTrain}
-          placeholder={trainsLoaded ? "Valitse juna" : "Ladataan..."}
-          optionFilterProp="children"
-          onSelect={(_value, option) => setSelectedTrain(option.value)}
-          onClear={() => setSelectedTrain(null)}
-          filterOption={(input, option) => {
-            if (!option) return false;
-
-            const terms = [
-              option.label,
-              option.arrivalStationName,
-              option.arrivalStationShortCode,
-              option.departureStationName,
-              option.departureStationShortCode,
-            ];
-
-            const keywords = input.split(" ");
-            if (keywords.length === 0) return false;
-            return keywords.every((keyword) =>
-              terms.some((term) => term.toLowerCase().includes(keyword.toLowerCase()))
-            );
-          }}
-          filterSort={(optionA, optionB) => parseInt(optionA.value) - parseInt(optionB.value)}
-          options={allTrains}
-        />
-
-        <br />
-
-        <br />
-
-        <Button
-          disabled={maintenance || (selectedDate && selectedTrain ? false : true)}
-          loading={trainLoading}
-          onClick={() => {
-            setTrainLoading(true);
-            router.push(`/train/${selectedDate}/${selectedTrain}`).catch(console.error);
-          }}
-        >
-          Jatka
-        </Button>
-
-        <div style={{ position: "absolute", bottom: "0", textAlign: "center" }}>
-          <Button
-            onClick={() => setIsFbModalOpen(true)}
-            style={{
-              fontWeight: 500,
-              height: "40px",
-              fontSize: "16px",
-              marginBottom: "10px",
-            }}
-          >
-            💬 Anna palautetta
-          </Button>
-
-          <p style={{ color: "#949090" }}>
-            Tämä on{" "}
-            <a
-              style={{ color: "#949090", textDecoration: "underline" }}
-              href="https://github.com/Chicken/VenaaRauhassa"
-            >
-              avoimen lähdekoodin
-            </a>{" "}
-            projekti
-          </p>
-
-          <a
-            style={{
-              color: "#B1B0B0",
-              fontSize: "10px",
-              marginLeft: "10px",
-              marginRight: "10px",
-              textDecoration: "underline",
-            }}
-            href="https://www.digitraffic.fi/kayttoehdot/"
-          >
-            Liikennetietojen lähde Fintraffic / digitraffic.fi, lisenssi CC 4.0 BY
-          </a>
-
-          <p
-            style={{
-              color: "#B1B0B0",
-              fontSize: "10px",
-              marginLeft: "10px",
-              marginRight: "10px",
-              marginTop: "5px",
-            }}
-          >
-            Emme ole <span style={{ fontStyle: "italic" }}>VR-Yhtymä Oyj:n</span>, sen
-            tytäryhtiöiden tai sen yhteistyökumppanien kanssa sidoksissa tai millään tavalla
-            virallisesti yhteydessä niihin. Virallinen verkkosivusto on osoitteessa{" "}
-            <a style={{ color: "#B1B0B0", textDecoration: "underline" }} href="https://www.vr.fi/">
-              www.vr.fi
-            </a>
-            .
-          </p>
-        </div>
+        <Footer setIsFbModalOpen={setIsFbModalOpen} />
       </Flex>
     </>
   );
